Move verification header to include
